Reorganise the code structure
Adds more space for implementing platform-independent thread priority routines:
1. Separates linux, unix, and all other (possible posix-compatible)
implementations into "unix" module.
2. Adds a separate module for windows-only routines.
3. Adds a few functions "set_current_thread_priority" and
"ThreadPriority::set_for_current" for better comform and for avoiding writing
repetitive code.
