+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "thread-priority"
-version = "0.1.1"
+version = "0.2.0"
 authors = ["Victor Polevoy <fx@thefx.co>"]
 description = "Library for managing threads priority and schedule policies"
 repository = "https://github.com/vityafx/thread-priority"
@@ -8,7 +8,12 @@ documentation = "https://docs.rs/thread-priority/"
 readme = "README.md"
 license = "MIT"
 keywords = ["thread", "schedule", "priority", "pthread"]
-categories = ["concurrency"]
+categories = ["concurrency", "asynchronous", "os"]
+edition = "2018"
 
-[dependencies]
+[target.'cfg(unix)'.dependencies]
 libc = "0.2"
+
+[target.'cfg(windows)'.dependencies]
+libc = "0.2"
+winapi = "0.3"
